requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
openai==1.77.0
duckduckgo-search==8.0.1
pydantic==2.11.4
python-dotenv==1.1.0
jinja2==3.1.6
httpx==0.28.1
tenacity==9.0.0
cachetools==5.5.0
trafilatura==1.12.2
